backend/requirements.txt
# ============================================================================
# Backend Dependencies - Planta de Tratamiento de Agua "La Esperanza"
# Consolidado: Backend + Machine Learning
# ============================================================================

# ===== Core Backend =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
alembic==1.12.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
openpyxl==3.1.5

# ===== Machine Learning =====
# Core ML Libraries
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Model Persistence
joblib==1.3.2

# Validation & Metrics
scipy==1.11.4

# Visualization (for notebooks and reports)
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Configuration & Utilities
pyyaml==6.0.1
python-dateutil==2.8.2
typing-extensions==4.9.0